refactor: split options.cpp to individual files
